+++ tests/embedding/test_logger.py
@@ -72,6 +72,17 @@ def test_logging_write_info_after_initialize_with_error_level(rootdir, pytestcon
     assert "0xdeadbeef" not in stderr
 
 
+@pytest.mark.parametrize("addin_configuration", ["Mechanical", "WorkBench", "Legacy"])
+@pytest.mark.embedding
+@pytest.mark.minimum_version(241)
+def test_addin_configuration(rootdir, pytestconfig, addin_configuration):
+    """Test that mechanical can start with both the Mechanical and WorkBench configuration."""
+    stderr = _run_embedding_log_test(
+        rootdir, pytestconfig, f"log_configuration_{addin_configuration}"
+    )
+    assert f"{addin_configuration} configuration!" in stderr
+
+
 @pytest.mark.embedding
 def test_logging_write_error_after_initialize_with_info_level(rootdir, pytestconfig):
     """Test that output is written when an error is logged when configured at the info level."""